+++ Carrot-Assistant/environment.yml
@@ -3,7 +3,7 @@ channels:
   - defaults
   - conda-forge
 dependencies:
-  - haystack-ai=2.1.1
+  - haystack-ai=2.3.1
   - huggingface_hub=0.22.2
   - python-dotenv=1.0.1
   - fastapi=0.110.3
